Made changes to enable structed PADS types to be computed fields. Changes included:   1. Removing an error message that prevented dynamic types as   computed fields.   2. Added a new function to the generated set:       int T_genPD(P_t *p, T* rep, T_pd *pd)      This function creates a parse descriptor based on the argument      rep. It uses where clauses, field predicates, etc. to set whether      there are any semantic errors.  It returns true if the rep is      error-free, false otherwise.
     Currently, the function assumes tht the argument pd has already
     been initialized to all zeros.  I'm not sure this is a good
     design choice.

     The function exists for all types except recursive types, which
     still need to be addressed.
  3. Changed the compilation of computed fields to use a copy function
     if one exists.
Needs more testing...

+++ pads/padsc/include/ptypes/Pchar.h
@@ -98,7 +98,7 @@ Perror_t Pchar_read(P_t *pads,const Pbase_m *m,Pbase_pd *pd,Pchar *c_out);
 
 ssize_t Pa_char_write2io(P_t *pads,Sfio_t *io,Pbase_pd *pd,Pchar *c);
 
-ssize_t Pa_char_write2buf(P_t *pads, Pbyte *buf, size_t buf_len, ssize_t *buf_full,
+ssize_t Pa_char_write2buf(P_t *pads, Pbyte *buf, size_t buf_len, int *buf_full,
                     Pbase_pd *pd, Pchar *c
                     );
 
@@ -107,7 +107,7 @@ ssize_t Pa_char_write_xml_2io(P_t *pads, Sfio_t *io, Pbase_pd *pd, Pchar *c,
                     );
 
 ssize_t Pa_char_write_xml_2buf(P_t *pads, Pbyte *buf, size_t buf_len,
-                    ssize_t *buf_full, Pbase_pd *pd, Pchar *c, const char *tag,
+                    int *buf_full, Pbase_pd *pd, Pchar *c, const char *tag,
                     int indent
                     );
